tapdb: add multiverse simple cache with global cache invalidation
In this commit, we add a simple cache for the multiverse. We'll cache
all the roots, and then use a map of LRU caches to cache the proofs for
a given root. When we insert something into the tree, we need to
invalidate the proof caches, as the merkle proof path has changed. We
also wipe the entire root cache as well since once we insert, the true
root isn't clear, so we'll wait for the next read to obtain that
information directly from the database. Purging ensures that we'll
eventually have a consistent cache.