tapdb: maintain in memory cache for universe sync configs
In this commits, we start to maintain an in memory cache for the
universe sync configs. As is, each time we go to query for or insert a
new proof, we need a db transaction to fetch the sync configs. This ends
up consuming a db connection and adds an extra round trip.

To bypass this, we start to maintain an in memory cache for the configs.