tapdb: add basic cache for the universe stats queries
In this commit, we add a basic cache for the universe stats queries. For
the main stats, we'll cache the result, then purge the cache in the
background to force fresh results. For the timeline query, we'll round
down to the nearest hour, then use an LRU cache to store the event
queries. Finally, for the asset sync stats, we'll combine these two
approaches, with an LRU cache that'll fully decay after the cache
interval.