universe: update MultiverseRoot to use a single db query
In this commit, we modify the `MultiverseRoot` method to use a single DB query. If we don't have anything to filter, then we'll fetch the global root for that proof type.

Otherwise, we'll query the DB for the leaves that match the set of IDs, then insert those into our in-memory tree.

This saves us N DB queries where N is the size of the set of Identifier.