rfq: convert to use fixed point ops, add property based tests
In this commit, we convert the existing logic to use purely fixed point
operations. ALong the way we arrive at a value of 11 for our internal
arithmetic scale value, meaning we compute with 11 places in the decimal
point, before converting back.

We also add some property based tests using rapid. The tests helps us to
bind the level of error/precision we're comfortable with with the
various conversion operations.