tapchannel: update createSweepVpackets to be second level aware
In this commit, we update `createSweepVpackets` to be second level aware. This means that if we detect an auxSigInfo, then we know the vPkt we need to create is actually just the second level vPkt. We also don't need to generate a new script key either.

Finally, we'll properly set the absolute delay if the sweep desc indicates as such.