rpc: add hodl invoice support
In this commit, we add hodl invoice support when creating invoice for TAP channels. The invoice creation proceeds as normal, but a special HodlInvoice field is set which contains a payment hash which must be set.

When invoices are created using this endpoint. They MUST be settled using the SettleInvoice RPC on the invoicesrpc server in lnd.