tapdb: update InsertScriptKey to allow flipping known to true
In this commit, we update `InsertScriptKey` to allow flipping known to true, if it was false before.

This is useful as at times a script key from a smart contract might have been inserting on disk, but with declared known as false. Then if we tried to insert it again, with known as true, the upsert logic would end up making no change on disk.